+++ libraries/chain/vm_api/permission.cpp
@@ -3,14 +3,14 @@
  *  @copyright defined in eos/LICENSE.txt
  */
 
-void unpack_provided_keys( flat_set<public_key_type>& keys, const char* pubkeys_data, size_t pubkeys_size ) {
+void unpack_provided_keys( flat_set<public_key_type>& keys, const char* pubkeys_data, uint32_t pubkeys_size ) {
    keys.clear();
    if( pubkeys_size == 0 ) return;
 
    keys = fc::raw::unpack<flat_set<public_key_type>>( pubkeys_data, pubkeys_size );
 }
 
-void unpack_provided_permissions( flat_set<permission_level>& permissions, const char* perms_data, size_t perms_size ) {
+void unpack_provided_permissions( flat_set<permission_level>& permissions, const char* perms_data, uint32_t perms_size ) {
    permissions.clear();
    if( perms_size == 0 ) return;
